requirements.txt
pymongo==4.1.1
python-dotenv==0.20.0

paho-mqtt~=1.6.1